[Performance] Improved performance by converting 4 render targets from RHI_Format_R16G16B16A16_Float to RHI_Format_R11G11B10_Float. Unnecessary alpha channel writes have also been removed from post processing passes.

+++ Data/shaders/Common_Buffer.hlsl
@@ -84,7 +84,7 @@ cbuffer BufferUber : register(b1)
     float g_mat_normal;
     float g_mat_height;
 
-    float g_mat_id;
+    uint g_mat_id;
     float g_is_transparent_pass;
     uint g_mip_count;
     uint g_work_group_count;
